+++ packages/emlint/src/rules.ts
@@ -259,6 +259,8 @@ import tagSpaceBeforeClosingBracket from "./rules/tag/tag-space-before-closing-b
 
 import tagSpaceBetweenSlashAndBracket from "./rules/tag/tag-space-between-slash-and-bracket";
 
+import tagTable from "./rules/tag/tag-table";
+
 import tagClosingBackslash from "./rules/tag/tag-closing-backslash";
 
 import tagVoidSlash from "./rules/tag/tag-void-slash";
@@ -1172,6 +1174,7 @@ defineLazyProp(
   "tag-space-between-slash-and-bracket",
   () => tagSpaceBetweenSlashAndBracket
 );
+defineLazyProp(builtInRules, "tag-table", () => tagTable);
 defineLazyProp(
   builtInRules,
   "tag-closing-backslash",
@@ -1977,7 +1980,7 @@ function normaliseRequestedRules(opts: RulesObj): RulesObj {
   }
 
   console.log(
-    `1980 normaliseRequestedRules() FINAL ${`\u001b[${33}m${`res`}\u001b[${39}m`} = ${JSON.stringify(
+    `1983 normaliseRequestedRules() FINAL ${`\u001b[${33}m${`res`}\u001b[${39}m`} = ${JSON.stringify(
       res,
       null,
       4